# Frontend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MosTransport Tram Load Forecast API
  version: 1.0.0
  description: API for AI forecast of tram route passenger load (hackathon).
servers:
  - url: /api/v1
paths:
...
  /forecast/recompute:
    post:
      operationId: recomputeForecast
      summary: Trigger forecast recompute (admin)
      tags: [admin]
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecomputeResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /telemetry:
    post:
      operationId: postTelemetry
      summary: Batch client telemetry
      tags: [telemetry]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryBatch'
      responses:
        '204':
          description: No Content
  /user-scenarios:
    get:
      operationId: listUserScenarios
      summary: List saved coefficient scenarios
      tags: [scenarios]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserScenarioList'
    post:
      operationId: createUserScenario
      summary: Save coefficient scenario
      tags: [scenarios]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserScenarioCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserScenario'
  /user-scenarios/{id}:
    delete:
      operationId: deleteUserScenario
      summary: Delete scenario
      tags: [scenarios]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
components:
...
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [message, request_id]
      properties:
        message:
          type: string
        request_id:
          type: string
        code:
          type: string
...
    Coefficients:
      type: object
      properties:
        k_weather:
          type: number
          default: 1
        k_event:
          type: number
          default: 1
        k_season:
          type: number
          default: 1
        k_traffic:
          type: number
          default: 1
...
    RecomputeResponse:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
        status:
          type: string
    TelemetryEvent:
      type: object
      required: [type, ts]
      properties:
        type:
          type: string
        ts:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    TelemetryBatch:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/TelemetryEvent'
    UserScenario:
      type: object
      required: [id, user_id, name, coefficients, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        name:
          type: string
        coefficients:
          $ref: '#/components/schemas/Coefficients'
        created_at:
          type: string
          format: date-time
    UserScenarioCreate:
      type: object
      required: [name, coefficients]
      properties:
        name:
          type: string
        coefficients:
          $ref: '#/components/schemas/Coefficients'
    UserScenarioList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserScenario'